Add updated Nvidia information
A recent Reddit post (linked below) details a simple fix for Nvidia cards not entering the loading screen of a game. This fix was manually tested on a Manjaro machine and it seemed to solve all issues even after multiple restarts, so it may be a good idea to include on top of the other fix. There also may be a way to automate this in the helper script, but I'm not familiar with the language enough to implement it effectively.

Source: https://www.reddit.com/r/leagueoflinux/comments/skyg70/fixing_the_nvidia_495_black_screen_crash_problem/